Made ticket editing field update use correct ticket info.
The component that created the ticket editing field was only being rendered once upon page load.
When its single-ticket parent updated due to state changes it did not update this form.
This was abnormal behavior compared to other components and was resistant to the normal tricks
for forcing it to re-render. As such, I added a workaround where the ticket path is used to
get the ticket information from the states tickets list.
This also required fixing the groupings and delete buttons.